AI-Helmet-Detection-Project/requirements.txt
opencv-python
tensorflow
numpy
Pillow
flask
flask-cors
werkzeug
matplotlib
seaborn
ultralytics
torch
torchvision
requests
python-dotenv
python-pptx
